Backend/requirements.txt
fastapi
request
scrapy